flowedit: Remove LoadedFlow, sync func edits, consolidate WindowState, NodeLayout refs (#2602)
1. Remove LoadedFlow wrapper — load_flow() returns FlowDefinition
   directly, using its existing lib_references/context_references fields.

2. Sync FunctionViewer edits to root_flow — after any mutation to
   viewer.func_def, sync the change to the canonical copy in
   root_flow.subprocesses via sync_func_def_to_root().

3. Consolidate all impl WindowState blocks into a single block in
   window_state.rs. Move methods from file_ops.rs, history.rs,
   initializer.rs, library_mgmt.rs, and main.rs. Move
   next_unique_io_name to utils.rs.

4. NodeLayout holds references (&ProcessReference, Option<&Process>)
   instead of clones, eliminating per-frame deep cloning of the
   subprocess tree. Remove Default impl. Update FlowCanvas and all
   tests.
