+++ flowedit/src/history.rs
@@ -7,13 +7,10 @@
 use std::path::{Path, PathBuf};
 
 use flowcore::model::connection::Connection;
-use flowcore::model::flow_definition::FlowDefinition;
 use flowcore::model::input::InputInitializer;
 use flowcore::model::name::Name;
 use flowcore::model::process::Process;
 use flowcore::model::process_reference::ProcessReference;
-
-use crate::WindowState;
 
 /// An editing action that can be undone and redone.
 #[derive(Debug, Clone)]
@@ -180,220 +177,11 @@ impl EditHistory {
     }
 }
 
-impl WindowState {
-    fn apply_undo(&mut self, flow_def: &mut FlowDefinition) {
-        if let Some(action) = self.history.undo() {
-            match action {
-                EditAction::MoveNode {
-                    index,
-                    old_x,
-                    old_y,
-                    ..
-                } => {
-                    if let Some(pref) = flow_def.process_refs.get_mut(index) {
-                        pref.x = Some(old_x);
-                        pref.y = Some(old_y);
-                    }
-                    self.status = String::from("Undo: move");
-                }
-                EditAction::ResizeNode {
-                    index,
-                    old_x,
-                    old_y,
-                    old_w,
-                    old_h,
-                    ..
-                } => {
-                    if let Some(pref) = flow_def.process_refs.get_mut(index) {
-                        pref.x = Some(old_x);
-                        pref.y = Some(old_y);
-                        pref.width = Some(old_w);
-                        pref.height = Some(old_h);
-                    }
-                    self.status = String::from("Undo: resize");
-                }
-                EditAction::CreateNode {
-                    index,
-                    ref process_ref,
-                    ..
-                } => {
-                    let alias = if process_ref.alias.is_empty() {
-                        crate::utils::derive_short_name(&process_ref.source)
-                    } else {
-                        process_ref.alias.clone()
-                    };
-                    if index < flow_def.process_refs.len() {
-                        flow_def.process_refs.remove(index);
-                        flow_def.subprocesses.remove(&alias);
-                        flow_def
-                            .connections
-                            .retain(|c| !crate::utils::connection_references_node(c, &alias));
-                    }
-                    self.status = String::from("Undo: create node");
-                }
-                EditAction::DeleteNode {
-                    index,
-                    process_ref,
-                    subprocess,
-                    removed_connections,
-                } => {
-                    flow_def.process_refs.insert(index, process_ref);
-                    if let Some((name, proc)) = subprocess {
-                        flow_def.subprocesses.insert(name, proc);
-                    }
-                    flow_def.connections.extend(removed_connections);
-                    self.status = String::from("Undo: delete node");
-                }
-                EditAction::CreateConnection { connection } => {
-                    let from_str = connection.from().to_string();
-                    let to_strs: Vec<String> =
-                        connection.to().iter().map(ToString::to_string).collect();
-                    flow_def.connections.retain(|c| {
-                        c.from().to_string() != from_str
-                            || c.to().iter().map(ToString::to_string).collect::<Vec<_>>() != to_strs
-                    });
-                    self.status = String::from("Undo: create connection");
-                }
-                EditAction::DeleteConnection { index, connection } => {
-                    flow_def.connections.insert(index, connection);
-                    self.status = String::from("Undo: delete connection");
-                }
-                EditAction::EditInitializer {
-                    node_index,
-                    ref port_name,
-                    ref old_init,
-                    ..
-                } => {
-                    self.apply_initializer_state(
-                        flow_def,
-                        node_index,
-                        port_name,
-                        old_init.as_ref(),
-                    );
-                    self.status = String::from("Undo: initializer");
-                }
-            }
-            self.canvas_state.request_redraw();
-            self.trigger_auto_fit_if_enabled();
-        }
-    }
-
-    fn apply_redo(&mut self, flow_def: &mut FlowDefinition) {
-        if let Some(action) = self.history.redo() {
-            match action {
-                EditAction::MoveNode {
-                    index,
-                    new_x,
-                    new_y,
-                    ..
-                } => {
-                    if let Some(pref) = flow_def.process_refs.get_mut(index) {
-                        pref.x = Some(new_x);
-                        pref.y = Some(new_y);
-                    }
-                    self.status = String::from("Redo: move");
-                }
-                EditAction::ResizeNode {
-                    index,
-                    new_x,
-                    new_y,
-                    new_w,
-                    new_h,
-                    ..
-                } => {
-                    if let Some(pref) = flow_def.process_refs.get_mut(index) {
-                        pref.x = Some(new_x);
-                        pref.y = Some(new_y);
-                        pref.width = Some(new_w);
-                        pref.height = Some(new_h);
-                    }
-                    self.status = String::from("Redo: resize");
-                }
-                EditAction::CreateNode {
-                    index,
-                    process_ref,
-                    subprocess,
-                } => {
-                    let idx = index.min(flow_def.process_refs.len());
-                    flow_def.process_refs.insert(idx, process_ref);
-                    if let Some((name, proc)) = subprocess {
-                        flow_def.subprocesses.insert(name, proc);
-                    }
-                    self.status = String::from("Redo: create node");
-                }
-                EditAction::DeleteNode {
-                    index,
-                    subprocess,
-                    removed_connections,
-                    ..
-                } => {
-                    if index < flow_def.process_refs.len() {
-                        let removed = flow_def.process_refs.remove(index);
-                        let alias = if removed.alias.is_empty() {
-                            crate::utils::derive_short_name(&removed.source)
-                        } else {
-                            removed.alias.clone()
-                        };
-                        flow_def.subprocesses.remove(&alias);
-                    }
-                    // Also remove re-inserted subprocess if it was restored during undo
-                    if let Some((ref name, _)) = subprocess {
-                        flow_def.subprocesses.remove(name);
-                    }
-                    for conn in &removed_connections {
-                        let from_str = conn.from().to_string();
-                        let to_strs: Vec<String> =
-                            conn.to().iter().map(ToString::to_string).collect();
-                        flow_def.connections.retain(|c| {
-                            c.from().to_string() != from_str
-                                || c.to().iter().map(ToString::to_string).collect::<Vec<_>>()
-                                    != to_strs
-                        });
-                    }
-                    self.status = String::from("Redo: delete node");
-                }
-                EditAction::CreateConnection { connection } => {
-                    flow_def.connections.push(connection);
-                    self.status = String::from("Redo: create connection");
-                }
-                EditAction::DeleteConnection { index, .. } => {
-                    if index < flow_def.connections.len() {
-                        flow_def.connections.remove(index);
-                    }
-                    self.status = String::from("Redo: delete connection");
-                }
-                EditAction::EditInitializer {
-                    node_index,
-                    ref port_name,
-                    ref new_init,
-                    ..
-                } => {
-                    self.apply_initializer_state(
-                        flow_def,
-                        node_index,
-                        port_name,
-                        new_init.as_ref(),
-                    );
-                    self.status = String::from("Redo: initializer");
-                }
-            }
-            self.canvas_state.request_redraw();
-            self.trigger_auto_fit_if_enabled();
-        }
-    }
-
-    pub(crate) fn handle_undo(&mut self, flow_def: &mut FlowDefinition) {
-        self.apply_undo(flow_def);
-    }
-
-    pub(crate) fn handle_redo(&mut self, flow_def: &mut FlowDefinition) {
-        self.apply_redo(flow_def);
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
+    use crate::WindowState;
+    use flowcore::model::flow_definition::FlowDefinition;
     use std::collections::BTreeMap;
 
     fn test_pref() -> ProcessReference {
