+++ flowedit/src/initializer.rs
@@ -1,121 +1,10 @@
-//! Initializer editor logic for applying and syncing initializer edits.
-
-use flowcore::model::flow_definition::FlowDefinition;
-use flowcore::model::input::InputInitializer;
-
-use crate::history::EditAction;
-use crate::utils::derive_short_name;
-use crate::{InitializerEditor, WindowState};
-
-impl WindowState {
-    pub(crate) fn apply_initializer_edit(
-        &mut self,
-        flow_def: &mut FlowDefinition,
-        editor: &InitializerEditor,
-    ) {
-        let alias = flow_def
-            .process_refs
-            .get(editor.node_index)
-            .map(|pr| {
-                if pr.alias.is_empty() {
-                    derive_short_name(&pr.source)
-                } else {
-                    pr.alias.clone()
-                }
-            })
-            .unwrap_or_default();
-
-        let old_init = flow_def
-            .process_refs
-            .get(editor.node_index)
-            .and_then(|pr| pr.initializations.get(&editor.port_name).cloned());
-
-        let new_init = match editor.init_type.as_str() {
-            "none" => None,
-            "once" | "always" => {
-                let value = serde_json::from_str(&editor.value_text)
-                    .unwrap_or_else(|_| serde_json::Value::String(editor.value_text.clone()));
-                let init = if editor.init_type == "once" {
-                    InputInitializer::Once(value)
-                } else {
-                    InputInitializer::Always(value)
-                };
-                Some(init)
-            }
-            _ => return,
-        };
-
-        let Some(pref) = flow_def.process_refs.get_mut(editor.node_index) else {
-            return;
-        };
-        match &new_init {
-            Some(init) => {
-                pref.initializations
-                    .insert(editor.port_name.clone(), init.clone());
-            }
-            None => {
-                pref.initializations.remove(&editor.port_name);
-            }
-        }
-
-        self.history.record(EditAction::EditInitializer {
-            node_index: editor.node_index,
-            port_name: editor.port_name.clone(),
-            old_init,
-            new_init,
-        });
-        self.canvas_state.request_redraw();
-        self.status = format!("Initializer updated on {}/{}", alias, editor.port_name);
-    }
-
-    #[allow(clippy::unused_self)]
-    pub(crate) fn apply_initializer_state(
-        &mut self,
-        flow_def: &mut FlowDefinition,
-        node_index: usize,
-        port_name: &str,
-        init: Option<&InputInitializer>,
-    ) {
-        if let Some(pref) = flow_def.process_refs.get_mut(node_index) {
-            match init {
-                Some(i) => {
-                    pref.initializations
-                        .insert(port_name.to_string(), i.clone());
-                }
-                None => {
-                    pref.initializations.remove(port_name);
-                }
-            }
-        }
-    }
-
-    pub(crate) fn handle_initializer_type_changed(&mut self, new_type: String) {
-        if let Some(ref mut editor) = self.initializer_editor {
-            editor.init_type = new_type;
-        }
-    }
-
-    pub(crate) fn handle_initializer_value_changed(&mut self, new_value: String) {
-        if let Some(ref mut editor) = self.initializer_editor {
-            editor.value_text = new_value;
-        }
-    }
-
-    pub(crate) fn handle_initializer_apply(&mut self, flow_def: &mut FlowDefinition) {
-        if let Some(editor) = self.initializer_editor.take() {
-            self.apply_initializer_edit(flow_def, &editor);
-        }
-    }
-
-    pub(crate) fn handle_initializer_cancel(&mut self) {
-        self.initializer_editor = None;
-    }
-}
+//! Initializer editor tests.
 
 #[cfg(test)]
 mod test {
-    use super::*;
+    use crate::{InitializerEditor, WindowState};
     use flowcore::model::flow_definition::FlowDefinition;
+    use flowcore::model::input::InputInitializer;
     use flowcore::model::name::Name;
     use flowcore::model::process_reference::ProcessReference;
     use std::collections::BTreeMap;
